+++ jest.config.js
@@ -2,4 +2,5 @@ module.exports = {
   preset: "ts-jest",
   setupFiles: ["./src/test.setup.ts"],
   collectCoverageFrom: ["./src/**"],
+  coverageReporters: ["html"],
 };