+++ typescript.novaextension/run.sh
@@ -4,17 +4,27 @@
 
 cd "$WORKSPACE_DIR"
 
-# note: --tsserver-path=".../tsserver.js" doesn't support debugging since it tries to fork and bind two processes to the same port
-
 # symlinks have issues when the extension is submitted to the library, so we don't use node_modules/.bin
-node \
-	"$INSTALL_DIR/node_modules/typescript-language-server/lib/cli.js" \
-	--stdio \
-	--tsserver-path="$TSLIB_PATH/tsserver.js"
 
-# use this for debugging
-# node \
-# 	--inspect-brk \
-# 	"$INSTALL_DIR/node_modules/typescript-language-server/lib/cli.js" \
-# 	--stdio \
-# 	--tsserver-path="$TSLIB_PATH/tsserver"
+if [ "$DEBUG" != "TRUE" ]
+then
+	node \
+		"$INSTALL_DIR/node_modules/typescript-language-server/lib/cli.js" \
+		--stdio \
+		--tsserver-path="$TSLIB_PATH/tsserver.js"
+else
+	if [ "$DEBUG_BREAK" ]
+	then
+		DEBUG_ARG="--inspect-brk"
+	else
+		DEBUG_ARG="--inspect"
+	fi
+	DEBUG_ARG="$DEBUG_ARG=127.0.0.1:$DEBUG_PORT"
+	# note: --tsserver-path=".../tsserver.js" doesn't support debugging since it
+	# tries to fork and bind two processes to the same port
+	node \
+		"$DEBUG_ARG" \
+		"$INSTALL_DIR/node_modules/typescript-language-server/lib/cli.js" \
+		--stdio \
+		--tsserver-path="$TSLIB_PATH/tsserver.js"
+fi